engine/backend/requirement.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
google-genai>=0.3.0
